[core/ui] Refactor Auction UI file structure in core/ui components
