+++ core/ui/src/components/Auction/AuctionForm/style.less
@@ -1,4 +1,4 @@
-@import '../../variable.less';
+@import '../../../variable.less';
 
 .candy-auction-form {
   max-width: 600px;
