+++ core/ui/src/components/Auction/AuctionModal/AuctionModalConfirmed.tsx
