+++ core/ui/src/components/Auction/AuctionModal/AuctionModalDetail.tsx
