+++ core/ui/src/components/Auction/AuctionModal/index.tsx
