+++ core/ui/src/components/Auction/AuctionModal/style.less
@@ -1,4 +1,4 @@
-@import '../../variable.less';
+@import '../../../variable.less';
 
 .candy-auction-modal-detail {
   width: 100%;
