+++ core/ui/src/components/Auction/AuctionNftHeader/index.tsx
