+++ core/ui/src/components/Auction/AuctionNftHeader/style.less
