+++ core/ui/src/components/Auction/CreateAuctionConfirm/style.less
@@ -1,4 +1,4 @@
-@import '../../variable.less';
+@import '../../../variable.less';
 
 .candy-auction-confirm-container {
   max-width: 600px;
